refactor - service class 위치 정정
